+++ src/test/java/uk/co/asepstrath/bank/TransactionTests.java
@@ -9,7 +9,7 @@ class TransactionTests {
 
     @Test
     void testGetters() {
-        Transaction transaction = new Transaction(new DateTime(2025, 2, 14, 8, 30, 0), 21, "8f95782c-7c83-4dd7-8856-0e19a0e0a075", "0043d8d9-846d-49cb-9b04-8d3823e9d8c9", "TOP", "PAYMENT");
+        Transaction transaction = new Transaction(new DateTime(2025, 2, 14, 8, 30, 0), BigDecimal.valueOf(21), "8f95782c-7c83-4dd7-8856-0e19a0e0a075", "0043d8d9-846d-49cb-9b04-8d3823e9d8c9", "TOP", "PAYMENT");
         assertEquals(new DateTime(2025, 2, 14, 8, 30, 0), transaction.getTimestamp());
         assertEquals(BigDecimal.valueOf(21), transaction.getAmount());
         assertEquals("8f95782c-7c83-4dd7-8856-0e19a0e0a075", transaction.getFrom());
